Refactor(InputView): `Validator` 생성 후 검증 로직 관리
